+++ src/examples.h
@@ -7,7 +7,7 @@
 
 void example_reverb() ;
 void example_chorus() ;
-void example_ks(int64_t start) ;
-void example_fm(int64_t start) ;
-void example_multimbral_fm(int64_t start) ;
-void example_drums(int64_t start, int loops);
+void example_ks(uint32_t start) ;
+void example_fm(uint32_t start) ;
+void example_multimbral_fm(uint32_t start) ;
+void example_drums(uint32_t start, int loops);
